+++ coolamqp/uplink/handshake.py
@@ -33,7 +33,7 @@ CLIENT_DATA = [
     (b'copyright', (b'Copyright (C) 2016-2024 SMOK sp. z o.o.', 'S')),
     (
         b'information', (
-            b'Licensed under the MIT License.\nSee https://github.com/smok-serwis/coolamqp for details',
+            b'Licensed under the MIT License.\nSee https://git.dms-serwis.com.pl/smokserwis/coolamqp for details',
             'S')),
     (b'capabilities',
      ([(capa, (True, 't')) for capa in SUPPORTED_EXTENSIONS], 'F')),