+++ CHANGELOG.md
@@ -9,3 +9,4 @@ Since v1.3.2 they'll be put here and in release description.
 * stress tests will run for 120 seconds now
 * stress tests will be harder, and use more queues
 * added arguments to queues, queue binds and exchanges
+* creating auto_delete non-exclusive queues will be met with a [PendingDeprecationWarning](https://www.rabbitmq.com/blog/2021/08/21/4.0-deprecation-announcements)